examples/langchain_example/requirements.txt
gati
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-core>=1.0.0
python-dotenv

